requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.22
sentence-transformers==2.3.1
pypdf==4.0.1
python-dotenv==1.0.1
gradio==4.16.0
openai==1.10.0